feat: change bookmark entity
