+++ src/main/java/com/kumarket/kumarket/bookmarks/BR.java
@@ -1,0 +1,11 @@
+//package com.kumarket.kumarket.bookmarks;
+//
+//import com.kumarket.kumarket.bookmarks.entities.BookmarkEntity;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//import java.util.List;
+//import java.util.Optional;
+//
+//public interface BookmarkRepository extends JpaRepository<BookmarkEntity, Long> {
+//    Optional<BookmarkEntity> findByUserId(Long id);
+//}
