+++ src/main/java/com/kumarket/kumarket/bookmarks/BookmarkRepository.java
@@ -4,8 +4,7 @@ import com.kumarket.kumarket.bookmarks.entities.BookmarkEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface BookmarkRepository extends JpaRepository<BookmarkEntity, Long> {
-    Optional<BookmarkEntity> findByUserId(Long id);
+    List<BookmarkEntity> findAllByUserId(Long id);
 }
