+++ src/main/java/com/kumarket/kumarket/bookmarks/entities/BookmarkEntity.java
@@ -1,14 +1,12 @@
 package com.kumarket.kumarket.bookmarks.entities;
 
+import com.kumarket.kumarket.posts.entities.PostEntity;
 import com.kumarket.kumarket.users.entities.UserEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "bookmark")
@@ -19,16 +17,20 @@ import java.util.List;
 public class BookmarkEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    private Long id;
 
-    @OneToOne
+    @ManyToOne
     @JoinColumn(name = "user_id")
-    UserEntity user;
+    private UserEntity user;
 
-    @OneToMany(mappedBy = "bookmark", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
-    private List<BookmarkPostEntity> bookmarkPost = new ArrayList<>();
+    @ManyToOne
+    @JoinColumn(name = "post_id")
+    private PostEntity post;
 
-    public void setUser(UserEntity user) {
-        this.user = user;
+    public void setUser(UserEntity userEntity) {
+        this.user = userEntity;
+    }
+    public void setPost(PostEntity postEntity){
+        this.post = postEntity;
     }
 }
