+++ src/main/java/com/kumarket/kumarket/posts/PostService.java
@@ -53,6 +53,7 @@ public class PostService {
                 .thumbnailUrl(postDto.getThumbnailUrl())
                 .photos(new ArrayList<>())
                 .viewCount(postEntity.getViewCount())
+                .user(postEntity.getUser())
                 .bookmarks(postEntity.getBookmarks())
                 .build();
         for(String url : postDto.getPhotos().getUrls()){
