+++ src/main/java/com/kumarket/kumarket/users/UserService.java
@@ -1,7 +1,5 @@
 package com.kumarket.kumarket.users;
 
-import com.kumarket.kumarket.bookmarks.BookmarkRepository;
-import com.kumarket.kumarket.bookmarks.entities.BookmarkEntity;
 import com.kumarket.kumarket.users.dto.UserDto;
 import com.kumarket.kumarket.users.entities.UserEntity;
 import lombok.RequiredArgsConstructor;
@@ -12,10 +10,10 @@ public class UserService {
     private final UserRepository userRepository;
 
     public void createUser(UserDto userDto){
-        BookmarkEntity bookmarkEntity = new BookmarkEntity();
+        //BookmarkEntity bookmarkEntity = new BookmarkEntity();
         UserEntity userEntity =userDto.toEntity();
-        bookmarkEntity.setUser(userEntity);
-        userEntity.setBookmark(bookmarkEntity);
+        //bookmarkEntity.setUser(userEntity);
+        //userEntity.setBookmark(bookmarkEntity);
         userRepository.save(userEntity);
     }
