+++ src/main/java/com/kumarket/kumarket/users/entities/UserEntity.java
@@ -1,7 +1,6 @@
 package com.kumarket.kumarket.users.entities;
 
-import com.kumarket.kumarket.bookmarks.entities.BookmarkEntity;
-import com.kumarket.kumarket.bookmarks.entities.BookmarkPostEntity;
+//import com.kumarket.kumarket.bookmarks.entities.BookmarkEntity;
 import com.kumarket.kumarket.common.BaseTimeEntity;
 import com.kumarket.kumarket.posts.entities.PostEntity;
 import jakarta.persistence.*;
@@ -32,21 +31,21 @@ public class UserEntity extends BaseTimeEntity {
 
     private Integer likes;
 
-    @OneToOne(cascade = CascadeType.ALL,mappedBy = "user")
-    BookmarkEntity bookmark;
+    //@OneToOne(cascade = CascadeType.ALL,mappedBy = "user")
+    //BookmarkEntity bookmark;
 
     @OneToMany(mappedBy = "user")
     private List<PostEntity> posts = new ArrayList<>();
 
 
     public void uploadPost(PostEntity postEntity){
-        //posts.add(postEntity);
+        posts.add(postEntity);
         postEntity.setUser(this);
     }
 
-    public void setBookmark(BookmarkEntity bookmark) {
-        this.bookmark = bookmark;
-    }
+//    public void setBookmark(BookmarkEntity bookmark) {
+//        this.bookmark = bookmark;
+//    }
     //    public void addBookmark(BookmarkPostEntity bookmarkPostEntity){
 //        bookmarks.add(bookmarkPostEntity);
 //        bookmarkPostEntity.setUser(this);
